Tree, details view and edit form
The list is now fully operational.
The showDetails is also implemented. When an item in the tree is selected, it gets displayed in the details view.
The details view also shows an Edit button. Pressing this will switch to the edit form. This shows the name and description of the category. It is currently not possible yet to actually save anything.